+++ server/dashboard.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/hacdias/eagle/v4/eagle"
 	"github.com/hacdias/eagle/v4/entry"
+	"github.com/hacdias/eagle/v4/renderer"
 	"github.com/hacdias/indieauth/v3"
 )
 
@@ -73,7 +73,7 @@ func (s *Server) dashboardPost(w http.ResponseWriter, r *http.Request) {
 }
 
 func (s *Server) serveDashboard(w http.ResponseWriter, r *http.Request, data interface{}) {
-	s.serveHTML(w, r, &eagle.RenderData{
+	s.serveHTML(w, r, &renderer.RenderData{
 		Entry: &entry.Entry{
 			FrontMatter: entry.FrontMatter{
 				Title: "Dashboard",
@@ -81,5 +81,5 @@ func (s *Server) serveDashboard(w http.ResponseWriter, r *http.Request, data int
 		},
 		Data:    data,
 		NoIndex: true,
-	}, []string{eagle.TemplateDashboard})
+	}, []string{renderer.TemplateDashboard})
 }
